requirements.txt
# Django related
Django==3.1.6
django-extensions==3.1.1
django-service-objects==0.7.0
djangorestframework==3.12.2
django-oauth-toolkit==1.5.0

# Code formatting
black==20.8b1
flake8==3.8.3
isort==5.7.0
pycodestyle==2.6.0

# misc
pillow==8.1.0